requirements.txt
Django==5.0.1
django-filter==23.5
drf-spectacular-sidecar==2024.1.1
drf-spectacular==0.27.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0